+++ lib/jxl/decode.cc
@@ -2798,7 +2798,7 @@ JxlDecoderStatus JxlDecoderGetBoxSizeRaw(const JxlDecoder* dec,
   return JXL_DEC_SUCCESS;
 }
 
-JxlDecoderStatus JxlDecoderGetBoxContentsSize(const JxlDecoder* dec,
+JxlDecoderStatus JxlDecoderGetBoxSizeContents(const JxlDecoder* dec,
                                               uint64_t* size) {
   if (!dec->box_event) {
     return JXL_API_ERROR("can only get box info after JXL_DEC_BOX event");